SDK: add SDK overlay for Windows
Introduce a WinSDK overlay for Windows.  This allows us to define some
shared constants that are not correctly imported right now.  This cleans
up the logic in the swift side of things and aids in the bring up.

Now that we have a SDK overlay for Windows, we should structure the tree
according to the OS family.