[cmake] Move swift_create_stdlib_targets from to AddSwiftStdlib.cmake.
This is part of my effort to split the stdlib build. Part of the work needed to
do this is we need to be able to given the same SWIFT_SDKS parameter come up
with the same list of targets to generate in the sub-stdlib cmake and in the
parent cmake (so that it can forward down those arguments to the sub-cmake for
the stdlib).