[cmake] Split the generation of stdlib-target names from creating the actual stdlib targets.
In order for the external stdlib split to be truly NFCI, we need to ensure that
when a user runs ninja ${swift-stdlib-target} the host cmake needs to forward
down that target to the child cmake.

In order to do this in a way that prevents a maintenance burden, I am splitting
the logic of /what/ stdlib targets to build from creating the actual
targets. Then in a subsequent commit I can reuse this logic to generate the list
of "forwarding" targets that the host cmake should create.