remote-run: make tests partially pass on Windows
Not all the tests are possible to run on Windows since they expect a
Unix-like shell environment.  However, the rest of the tests can be
accommodated.  This actually found an issue in the implementation.
Since the implementation assumes the target is POSIX/Unix-y, we should
use posixpath rather than os.path which uses the host's path style.