Update Windows.md
Update the cmake configuration to actually reflect the current state of the world.  This now builds against the MSVC built LLVM/clang and builds with clang-cl.  This also enables building the test suite since that is starting to work on Windows.  The MSVC commands are still out of date, but this is an overall improvement to the instructions.  Since the primary work shifted towards Windows x86_64, we can use the prebuilt binaries for ICU from the ICU project as they provide Windows x86_64 builds.