[Parse] Diagnose duplicate @available arguments
Previously, if the `message` or `available` argument to an `@available` attribute was specified two or more times, the last value would be used and the earlier ones would be silently discarded. This change emits a warning instead; we can strengthen this into an error in a future version of Swift.

Fixes rdar://problem/46348825.