put autodiff assoc fns in witness table (#20759)
The overall effect of this PR is to put autodiff_associated_function entries in the witness table for protocol function requirements that are marked @differentiable. See test/AutoDiff/witness_table.swift for a detailed example.

Here's a summary of the changes involved:

AutoDiff changes
Added a new AutoDiffAssociatedFunctionIdentifier type to act as an index for these new entries.
Fixed AnyFunctionType::getAutoDiffAssociatedFunctionType to work when parameters/results are generic (conforming to Differentiable).
Added selfUncurried flag to various functions. See the comments for explanation. This is used in thunk generation.
SILFunctionType::getAutoDiffAssociatedFunctionType: taught it about parameterIndices and added isMethod flag. This stuff is used in thunk generation.

SILWitnessTable changes
Added a new AutoDiffAssociatedFunction kind for this new kind of entry.
Added printing and parsing for the new kind of entry.

SILWitnessVisitor changes
Made it call a new visitor method addAutoDiffAssociatedFunction when it sees a function requirement with @differentiable.

SILGenWitnessTable and witness thunk generation changes
Taught SILGenWitnessTable how to addAutoDiffAssociatedFunction.
Added a new AutoDiffAssociatedFunctionIdentifier *autoDiffFuncId parameter to a bunch of thunk-generating functions. When it's nullptr, the function does the usual thing. When it's non-null, the function generates a thunk for the autodiff associated function instead. I think it would be cleaner to have separate functions for these things. That would involve some refactoring, and I think it would be easiest to defer this until we upstream these changes to master because I might have to throw away any refactoring that I do now if it conflicts with a lot of changes in master.
IRGen
Added new kind of WitnessTableEntry for this new kind of entry.
Taught WitnessTableLayout and WitnessTableBuilder how to addAutoDiffAssociatedFunction