ABI: use elaborated type for type reference (NFC)
Use the elaborated type for the type alias that we are creating.
Because the type inherits from `TargetMetadata`, the reference here is
parsed as a reference to the underlying type in MSVC.  Use the
elaborated type to resolve to the type itself.  The rules for these
changed around C++11, but MSVC defaults to the old style of the name
resolution.  NFC.