synthesize vjp (#20980)
This PR synthesizes the VJP.

Importantly, I create the VJP immediately when the task is registered, so that PrimalGen will be able to reference the VJP. This required me to also create the empty primal and adjoint immediately when the task is registered, so that the VJP can reference the primal and adjoint. This means that we can no longer check if primal/adjoint needs to be synthesized by checking if task.primal/task.adjoint is nullptr, so I created a new piece of state FunctionSynthesisState to keep track of that.

The VJP synthesis itself, in DifferentiationTask::createVJP is a pretty straightforward imitation of the existing gradient synthesis code.

Miscellaneous changes required to make it all work:
*A lot of the tests differentiate functions with args/results that do not conform to Differentiable. This breaks getAutoDiffAssociatedFunctionType(), which createVJP() calls. So I added a quick workaround to make getAutoDiffAssociatedFunctionType() treat types as their own Tangent/Cotangent when they do not conform to Differentiable.
* getAutoDiffAssociatedFunctionType() needed a resultIndex to make some tests pass.
* Rearranged the order of some sil functions in some tests because now they get generated in a different order.

I added on SIL test for the synthesized VJP. Once I make the PR that uses the VJP, I should be able to test it more thoroughly by actually executing it.