Generalize ConcreteExistentialInfo.
Handle the unconditional_checked_cast case.

This involved a major cleanup/rewrite of the utility for analyzing
existential types.

Rewrite potentially fixes some latent bugs in the
ExistentialSpecializer that I found by code inspection.

There still appears to be a bug where it bit-casts an opened
existential to a concrete type, but I believe that code path is
disabled.