[Type checker] Completely ignore SIMD operators in the first type-checking pass
Even the presence of the SIMD operators can cause the type checker to exhaust
it's search space. Therefore, implement a two-stage hack to completely
ignore the SIMD operators and then, if we saw both SIMD types and SIMD
operators in the expression, make a second pass through the type
checker, enabling the SIMD operators as a fallback.

As with the previous hack, which this builds on, it's disabled when
the operator designated types functionality is enabled, because that
change fixes the problem.

Fixes the *other* regressions we saw with rdar://problem/46541800.

(cherry picked from commit 9772e89b8a2276ecffc1b91a4aadcb17f13c14cb)