differentiate witness methods (#21286)
The main idea is the new emitVJPReference function, which does the existing lookup logic for function_ref, and which does new lookup logic for witness_method. I refactored the 2 places that had function_ref lookup logic (PrimalGen::visitApplyInstWithVJP and Differentiation::processAutoDiffFunctionInst) to call emitVJPReference.

I added a SIL FileCheck test for it. An end-to-end test is currently not possible: It requires the builtins for looking up method VJPs. I will work on that next in a separate PR.

Other minor necessary fixes:
* Made SILFunctionType::getAutoDiffAssociatedFunctionType work on @convention(witness_method) functions.
* Made differentiation process transparent functions. The witness thunks with autodiff_functions are transparent for some reason. See test/AutoDiff/witness_table_silgen.swift for an example. The CHECKs in that are checking witness thunks.
* Fixed bug so that differentiation processes autodiff_functions in modules without #gradient or [differentiable].
* And finally, I re-enabled the SILVerifier check because now it passes.