docs: running tests on Windows
Start documenting how to run the test suite on Windows so that others can easily reproduce the test results.  Document the external dependencies from the test suite.  Since the LLVM test suite depends on the GNUWin32 tools, I believe that this dependency is not too onerous.  Over time, it may be possible to implement more of the functionality in lit to reduce the dependency, but, for now, use the tools to speed up the test suite bring up.