[Name lookup] Narrow shadowing to non-member types.
Narrow the recently-introduced shadowing rule for types so that it only
applies to non-member types. Member types lack a reasonable syntax for
specifying precisely which module to look into, and there are a few use
cases where the type checker will pick a type that would be shadowed by
the new rule.

Fixes a source-compatibility regression introduced by the shadowing rule.